requirements.txt
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.1.0
Pillow==5.3.0
PyJWT==1.6.4
PyMySQL==0.9.3
PyYAML==3.13
Pygments==2.2.0
cached-property==1.5.1
certifi	2018.10.15
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
diff-match-patch==20181111
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-filter==2.0.0
django-formtools==2.1
django-guardian==1.4.9
django-import-export==1.1.0
django-redis==4.10.0
django-reversion==3.0.2
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
docker==3.7.0
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
drf-extensions==0.4.0
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.9.2
idna==2.7
itypes==1.1.0
jdcal==1.4
jsonschema==2.6.0
mysqlclient==1.3.13
oauthlib==3.0.1
odfpy==1.3.6
openpyxl==2.5.11
pip==19.0.2
pycryptodome==3.7.3
python3-openid==3.1.0
pytz==2018.7
redis==3.1.0
requests==2.20.1
requests-oauthlib==1.2.0
setuptools==40.8.0
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==3.0.0
tablib==0.12.1
texttable==0.9.1
uWSGI==2.0.18
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24.1
websocket-client==0.54.0
wheel==0.33.0
xlrd==1.1.0
xlwt==1.3.0